Fixed newline issue in titles
Previously, if you were to put '\n' in your title to make a newline, it would not translate to the tex code, and instead would be eaten by the transformation. Now, it should add \\ instead, which won't be eaten, and add  "title style={align=center}" which is needed for the \\ to work. Could also change this to be more customizable (e.g. align=left, align=right) if needed.